Crate tink_daead

This crate provides implementations of the tink::DeterministicAead primitive.

Unlike AEAD, implementations of this interface are not semantically secure, because encrypting the same plaintex always yields the same ciphertext.

Modules

subtle
Provides subtle implementations of the DeterministicAEAD primitive using AES-SIV.

Constants

AES_SIV_KEY_VERSION
Maximal version of AES-SIV keys.

AES_SIV_TYPE_URL
Type URL of AES-SIV keys that Tink supports.

UPSTREAM_VERSION
The upstream Tink version that this Rust port is based on.

Functions

aes_siv_key_template
Return a KeyTemplate that generates a AES-SIV key.

init
Initialize the tink-daead crate, registering its primitives so they are available via tink.

new
Return a tink::DeterministicAead primitive from the given keyset handle.

new_with_key_manager
Return a tink::DeterministicAead primitive from the given keyset handle and custom key manager.